pyproject.toml
[tool.poetry]
name = "Xi_Lusheng"
version = "0.1.0"
description = "Xi_Lusheng"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = [
    "nonebot_plugin_gocqhttp",
    "nonebot_plugin_ygo",
    "nonebot_plugin_repeater",
    "nonebot_plugin_remake",
    "nonebot_plugin_baidutranslate",
    "nonebot_plugin_abstract",
    "nonebot_plugin_crazy_thursday",
    "nonebot_plugin_gspanel",
    "nonebot_plugin_gsmaterial",
    "nonebot_plugin_PicMenu"
]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
